+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.2.0 [03/05/2019]
+### Additions
+- User under repository name is now a link to their profile page.
+
 ### Changes
 - Made the page more mobile friendly.
+- Made the View on GitHub button outlined.
 
 ## 1.1.0 [03/04/2019]
 ### Changes